Add ability to specify server version on server change
When changing a server the new endpoint can be specified in the kickmessage, e.g. "127.0.0.1:25566;40". The last field is optional. If it is not present (or 0) the server will give plugins the possibility to set the version number. if no plugin set it the server look into the configuration file. If there is a entry for the specific end poin this version number is used. if there is not such a entry it will fall back to use the last possible server version the proxy server can handle.

I hope this fixes #19